3 model development/structure/codes/requirements.txt
pytorch
pandas
transformers
numpy
tqdm
re
datetime
random
joblib
scikit-learn
matplotlib
sentence_transformers
xgboost

